requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.2.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
distro==1.9.0
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.114.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
jiter==0.5.0
lxml==5.3.0
openai==1.45.0
paramiko==3.4.1
passlib==1.7.4
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.4.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
pymongo==4.8.0
PyNaCl==1.5.0
PyPDF2==3.0.1
pypdfium2==4.30.0
python-docx==1.1.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
rsa==4.9
scp==0.15.0
six==1.16.0
sniffio==1.3.1
starlette==0.38.5
tqdm==4.66.5
typing_extensions==4.12.2
uvicorn==0.30.6
